changed poly3d verts to doubles, fixed user input rotate, translate, scale methods to use world methods

+++ LabSoftware VS2010/GraphicsUtil.h
@@ -11,7 +11,7 @@ typedef struct  {int x, y;} POINT2D;
 
 typedef struct { int x; int y; RGBColour* c; } VERTEX;
 
-typedef struct { int x; int y; int z; RGBColour* c; } VERTEX_3D;
+typedef struct { double x; double y; double z; RGBColour* c; } VERTEX_3D;
 
 typedef struct { double x; double y; double z; } VERTEX_3D_f;
